+++ Project/include/ray.hpp
@@ -12,8 +12,8 @@
  */
 struct Ray
 {
-    Vec3d origin = {0, 0, 0};
-    Vec3d direction = {1, 0, 0};
+    Vec3f origin = {0, 0, 0};
+    Vec3f direction = {1, 0, 0};
 
     /**
       * @brief Default construct a ray with origin at <0,0,0> and facing 0 direciton
@@ -24,12 +24,12 @@ struct Ray
      * @brief Construct a ray by its origin and direction
      * @related Ray
      */
-    constexpr Ray(Vec3d a, Vec3d b) : origin{a}, direction{b} {}
+    constexpr Ray(Vec3f a, Vec3f b) : origin{a}, direction{b} {}
 
     /**
      * @brief Gets the result point after we put the parameter t into the ray function
      */
-    constexpr Vec3d point_at_parameter(double t) const noexcept {
+    constexpr Vec3f point_at_parameter(float t) const noexcept {
         assert(t >= 0);
         return origin + t * direction;
     }
